added async two-stage chart marker story
