+++ src/map/config/map.json
@@ -43,6 +43,14 @@
     "#593315",
     "#F13A13",
     "#232C16" 
+  ],
+
+  "chartMarkerColorsHex": [
+    "#0018A9",
+    "#3B1988",
+    "#771A66",
+    "#B21B45",
+    "#ED1C23"
   ]
 }
